# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для работы с отелями, комнатами, отзывами и избранным",
        "title": "StayGo Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/hotels/{hotelid}/rooms": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Получить список комнат по ID отеля",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID отеля",
                        "name": "hotelid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Room"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid hotel id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "failed to get rooms",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/rooms": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Создать комнату",
                "parameters": [
                    {
                        "description": "Данные комнаты",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Room"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Room"
                        }
                    },
                    "400": {
                        "description": "invalid parameters | invalid input | invalid hotel_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "access denied",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rooms/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Поиск комнат по городу, гостям и датам",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Город",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Количество гостей",
                        "name": "guests",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата заезда (YYYY-MM-DD)",
                        "name": "checkin",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата выезда (YYYY-MM-DD)",
                        "name": "checkout",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Room"
                            }
                        }
                    },
                    "400": {
                        "description": "city and guests are required | invalid guests | invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/rooms/{roomid}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rooms"
                ],
                "summary": "Получить комнату по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID комнаты",
                        "name": "roomid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Room"
                        }
                    },
                    "400": {
                        "description": "invalid room id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "room not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Room": {
            "description": "Информация о комнате, включая вместимость, цену, рейтинг и привязку к отелю",
            "type": "object",
            "properties": {
                "beds": {
                    "description": "Количество спальных мест",
                    "type": "integer",
                    "example": 2
                },
                "description": {
                    "description": "Описание комнаты",
                    "type": "string",
                    "example": "Уютный номер с видом на город"
                },
                "hotel_id": {
                    "description": "Идентификатор отеля, к которому относится комната",
                    "type": "integer",
                    "example": 101
                },
                "id": {
                    "description": "Уникальный идентификатор комнаты",
                    "type": "integer",
                    "example": 2001
                },
                "price": {
                    "description": "Цена за ночь (в целых единицах, например, рублях)",
                    "type": "integer",
                    "example": 4500
                },
                "rating": {
                    "description": "Текущий рейтинг комнаты (1-5)",
                    "type": "integer",
                    "example": 4
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}